requirements.txt
synapse-snappy-lib
paho-mqtt
tornado
snapstraction
